[up]微修正

+++ Source/BasePlayer.h
@@ -68,15 +68,15 @@ public:
 	 virtual void All();            //プレイヤーの処理
 	 virtual void Attack();         //攻撃処理
 
-	 virtual void Set_x(int _x) { x = _x; }                        //セッター
-	 virtual void Set_y(int _y) { y = _y; }                        //セッター
-	 virtual void Set_width(int _width) { width = _width; }        //セッター
-	 virtual void Set_height(int _height) { height = _height; }    //セッター
+	 virtual void Set_x(float _x) { x = _x; }                        //セッター
+	 virtual void Set_y(float _y) { y = _y; }                        //セッター
+	 virtual void Set_width(float _width) { width = _width; }        //セッター
+	 virtual void Set_height(float _height) { height = _height; }    //セッター
 
-	 virtual int Get_x() { return x; }                             //ゲッター
-	 virtual int Get_y() { return y; }                             //ゲッター
-	 virtual int Get_width() { return width; }                     //ゲッター
-	 virtual int Get_height() { return height; }                   //ゲッター
+	 virtual float Get_x() { return x; }                             //ゲッター
+	 virtual float Get_y() { return y; }                             //ゲッター
+	 virtual float Get_width() { return width; }                     //ゲッター
+	 virtual float Get_height() { return height; }                   //ゲッター
 
 };
